(#25) Remove hardcoded secret key from settings.py
settings.py now reads the secret key from the environment variables, so
`export SECRET_KEY="secretkey"` must be executed before running the
server. If this is not set, a django ImproperlyConfigured exception will
be raised. Note that there is a secret key hardcoded but only used if
DEBUG=True and `SECRET_KEY` is not set